Move pool tests into their own test and disable for race
